feat: notifications/support modules, Lambda test coverage, stop tracking vendored deps
- New backend modules: Notification + SupportTicket (DynamoDB-backed
  entities/repositories, use cases, mappers, controllers), device-token
  registration, plus Flyway migrations V3-V5 (notifications, support
  tickets, transaction idempotency keys).
- Fixed kycApi.verify()/usersApi.deleteAccount() sending JSON bodies to
  Spring endpoints declared with @RequestParam (Spring doesn't bind
  @RequestParam from a JSON body) - both were silently 400ing; switched
  the frontend calls to send query params instead.
- Added missing frontend wrappers with real UI for 3 backend endpoints
  that had none: POST /api/auth/resend-code (with a 60s cooldown),
  DELETE /api/users/account (danger-zone card + password-confirm
  modal), DELETE /api/kyc/documents/{id} (delete button, gated off
  APPROVED documents).
- Full unit/integration test coverage added across the backend, every
  Lambda, and the frontend api/lib layer (previously only 10 tests
  existed).
- .gitignore was accidentally full of pasted shell commands instead of
  ignore rules, so nothing was actually being filtered: stopped
  tracking ~2100 vendored Python packages (boto3/botocore et al.) that
  had been committed under neobank-lambdas/fraud-checker/, plus Lambda
  deploy zips and __pycache__/.venv across every Lambda; real ignore
  rules now cover Python/Maven/Node build output, secrets, and
  Terraform state.
- Maven wrapper (mvnw/.mvn) added so the backend builds without a
  system-installed Maven; Dockerfile per Lambda; CI workflows; JMeter
  load tests; Postman collection.

+++ neobank-lambdas/transaction-service/src/main/java/neobank/TransactionResponse.java
@@ -3,6 +3,8 @@ package neobank;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public class TransactionResponse {
 
     private String status;
@@ -16,17 +18,17 @@ public class TransactionResponse {
     @JsonProperty("to_account")
     private String toAccount;
 
-    private Double amount;
+    private BigDecimal amount;
 
     @JsonProperty("new_balance")
-    private Double newBalance;
+    private BigDecimal newBalance;
 
     private Long timestamp;
 
     public TransactionResponse() {}
 
     public TransactionResponse(String status, String transactionId, String fromAccount,
-                               String toAccount, Double amount, Double newBalance) {
+                               String toAccount, BigDecimal amount, BigDecimal newBalance) {
         this.status = status;
         this.transactionId = transactionId;
         this.fromAccount = fromAccount;
@@ -48,11 +50,11 @@ public class TransactionResponse {
     public String getToAccount() { return toAccount; }
     public void setToAccount(String toAccount) { this.toAccount = toAccount; }
 
-    public Double getAmount() { return amount; }
-    public void setAmount(Double amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
 
-    public Double getNewBalance() { return newBalance; }
-    public void setNewBalance(Double newBalance) { this.newBalance = newBalance; }
+    public BigDecimal getNewBalance() { return newBalance; }
+    public void setNewBalance(BigDecimal newBalance) { this.newBalance = newBalance; }
 
     public Long getTimestamp() { return timestamp; }
     public void setTimestamp(Long timestamp) { this.timestamp = timestamp; }
